+++ tools/cathedral/src/EventPromptUI.cpp
@@ -75,7 +75,7 @@ void EventPrompt::Load(const std::shared_ptr<objects::Event>& e)
         return;
     }
 
-    prop->message->SetValue(mEvent->GetMessageID());
+    prop->message->SetValue((uint32_t)mEvent->GetMessageID());
 
     for(auto choice : mEvent->GetChoices())
     {
@@ -92,7 +92,7 @@ std::shared_ptr<objects::Event> EventPrompt::Save() const
 
     Event::Save();
 
-    mEvent->SetMessageID(prop->message->GetValue());
+    mEvent->SetMessageID((int32_t)prop->message->GetValue());
 
     auto choices = prop->choices->GetObjectList<objects::EventChoice>();
     mEvent->SetChoices(choices);
